requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.8
django-widget-tweaks==1.4.1
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
static3==0.7.0
urllib3==1.22
django-braces==1.12.0
django-model-utils==3.1.1
django-star-ratings==0.6.1
swapper==1.1.0
django-storages==1.6.5
boto3==1.5.31
botocore==1.8.45
docutils==0.14
jmespath==0.9.3
python-dateutil==2.6.1
s3transfer==0.1.13